+++ CHANGELOG.md
@@ -9,3 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `slice` filter
 - `classnames` shortcode
 - `styles` shortcode
+- Shortcode options support
+- `includeGlob` shortcode